Fix the clients stuck, due to xmlrpc service fail
Clients stuck at the end of the service restarted.

Find the way to handle the xmlrpc client for multiple network,
which in the first time design to be it's virtual server ip
but the problem is the limitation of gnet_xmlrpc_client_new()
that can only parse the "localhost" as the hostname.
Therefore try to workaround by create the alias in the
apache config to do the special /rahunas_service path to
serve the xmlrpc stop accouting request.

2009-04-03  Neutron Soutmun <neo.neutron@gmail.com>

	* src/rh-config.{h,c}, src/rh-server.c , src/rh-xmlrpc-cmd.c:
	  - Drop all xml_serv_* due to now only the "localhost" that
	    gnet_xmlrpc_client_new() can make a connection to.
	  - Substitute xml_serv_* with the XMLSERVICE_HOST, XMLSERVICE_URL and
	    XMLSERVICE_PORT.
	  - According to XMLSERVICE_URL  "/rahunas_service/xmlrpc_service.php",
	    needs the special apache config to do the alias for this path,
	    will do in the debian apache config file for example.